RF(first wave): .distribution -> .distributions, .provenance -> .formats
