+++ niceman/formats/tests/test_provenance.py
@@ -10,7 +10,7 @@
 import logging
 
 from niceman.utils import swallow_logs
-from niceman.provenance import Provenance
+from niceman.formats import Provenance
 
 import niceman.tests.fixtures
