+++ niceman/retrace/__init__.py
@@ -6,7 +6,7 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-"""Determine package and distribution information for files
+"""Determine package and distributions information for files
 
 """
